Add the grouped-execution GPU benchmark and align its CE with train_model
grouped_loss_backward upcast logits with an explicit .float() before cross
entropy; train_model's padded path does not. Autocast's fp32 policy already
promotes cross_entropy - verified directly, bf16 logits in, fp32 loss out - so
the cast changed no numbers but did make the two paths allocate differently,
which would have shown up as a memory difference attributed to grouping rather
than to the cast. Removed so the A/B compares execution only.

The benchmark reports compile time and steady-state time separately and never
sums them: a one-off compile is amortized across millions of steps, while a
per-step recompile is not. After warmup it sets error_on_recompile and runs
further steps, so continued recompilation is caught explicitly rather than
being inferred from timings - and exceeding dynamo's recompile limit falls back
to eager silently, which would otherwise look like an unexplained slowdown.

Also records peak allocated memory and the real (B, T) subgroup distribution,
because whether recompilation converges depends on how many distinct shapes
real batches actually produce.

+++ src/exp0/grouped_execution.py
@@ -122,8 +122,12 @@ def grouped_loss_backward(
             logits = forward_fn(sub_tuples, sub_targets)
             # reduction="sum" then a single global divide: this is what keeps
             # the objective token-weighted instead of group-weighted.
+            # No explicit .float(): the padded path in train_model does not cast
+            # either, and autocast's fp32 policy already promotes cross_entropy,
+            # so casting here would only make the two paths differ in memory
+            # behaviour while producing the same numbers.
             loss_sum = F.cross_entropy(
-                logits.reshape(-1, logits.size(-1)).float(),
+                logits.reshape(-1, logits.size(-1)),
                 sub_mask[:, 1:].reshape(-1),
                 ignore_index=IGNORE_INDEX,
                 reduction="sum",
